Spent all damn day on this thing. There are many changes.
One, renamed the `templates` directory to `local`, along with all the modules in there.

Two, added a `Reference` base field type, and two subtypes: `LocalFile` and `Entry`. A `LocalFile` is a reference to a local file. An `Entry` is a reference to a content file somewhere in in the `local/content` directory tree. All specs that can be used in content files (e.g., in `local/specs/content`) can now include additional methods: `content_path`, `public_path`, `view_file`, etc. If these methods are defined in the spec (as module methods, not instance methods, for reasons explained later), then they will be pulled in and defined on the object the spec creates. If not, then the spec's `type` must define them. These methods are defined on the module because the `Entry` field type won't want to instantiate the tempate object just to verify that the slug its given is valid, but will need to validate that slug before setting it on the object. The `Entry` will instantiate the template it refers to via its `resolve` method.

Three, added an `Extendable` mixin, to allow for easily and uniformly extending an object in the manner that the Entry made necessary.

Four, all Field and Template objects are not instantiated with the spec they're bulding from. This allows for uniform `extend`ind, etc.

Five, the `Asset` field is now a `LocalFile` field.

There might be other things as well. This is really poor form. But the cms is in a pretty good state now!