Revise the way views are specified.
When rendering the view of some content, it'd be nice to be able to render it in different ways -- HTML, RSS, whatever you want. That's a feature that this commit starts to introduce.

Now, the `to_view` method of Renderables and the `view_file` method of Specs must accept a symbol parameter. This symbol will indicate the type of view that the procedure wants to render. This allows each Spec to specify a `view_file` for each desired type.

The directory structure can (should) have different trees for each render type, like
local/views/
    html/
    rss/