Revise the way markdown files can be rendered.
For content files that combine YAML and markdown, the default reader (File.read, piped to YAML.load) can't be used, a special (spec-specific) reader is needed. This commit introduces functionality for handling that.

Also, for content files that don't specify a `spec`, and also because it could be useful for testing new specs, a spec can be associated with a content file on the command line.