A problem came up when working on the PagesByTag Group.
The return from PagesByTag.prepare is an array of TaggedGroup Templates. The value of that spec's `group` field is a Group. In `prepare`, the Groups are being made. However, the `::Base::Fields::Group.make` method expects to make the group `from_spec`.

This problem might already be fixed in another branch.